Finished Animation through tex coords and mix value.
tex coords not finished because i realisede a better way to do things, Make structs for triangles and quads that hold their VAO EBO and stuff as well as a dynamic vector for textures and vertices.

struct Quad {
    GLuint program;
    GLuint vao;
    GLuint vbo;
    GLuint ebo;
    std::vector<GLuint> textures;
    std::vector<GLfloat> vertices;
};

now make the init method better for setting things up and pass out a vector of structs.
Even better:
1. Define a Base Class
Create a base class Shape with virtual methods for initialization and rendering.

2. Define Triangle and Quad Structs
Derive Triangle and Quad structs from the Shape base class.
Give each their own render and init method.

    Init to set up vertex data, Gen and bind VAO, VBO EBO, and glVertexAttribPointer stuff

       void render() override {
        glBindVertexArray(vao);
        glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
        glBindVertexArray(0);
    }

3. Initialize and Render Shapes
Create a method to initialize and render a list of shapes.

init holds a vector of shapes, callls init method for each shape and adds them to vector, then returns shapes vector.

render takes in the vector of shapes, iterates it and calls each shapes render function. (can then just chuck this in the main render() )